src/requirements.txt
django==1.2.7
pillow
qiniu==6.1.4